Condensed methods/classes a bit
