+++ app/src/main/java/com/example/brianduffy/masevo/Event.java
@@ -15,12 +15,24 @@ import java.util.HashSet;
 public abstract class Event implements Serializable{
     //ChatRoom chatRoom = new ChatRoom();
     // Need to implement ChatRoom
-    public Date startTime;
+    public int eventID;
+        // int storing the ID of the event.
+    public String eventName;
+        // String storing the display name of this event
+    public String eventDesc;
+        // Description of the event
+    public Date startDate;
         // Start time of the event
-    public Date endTime;
+    public Date endDate;
         // End time of the event
     public Location location;
         // Location of the event
+    public float radius;
+        // Radius of the event as set by a host
+    //public MapView curEventMapView
+    // Reference to the map for this event
+    public String hostEmail;
+        //The host's email
     public HashSet<String> hostList = new HashSet<>();
         // HashSet of hosts for the event
     public HashSet<String> attendeeList = new HashSet<>();
@@ -29,15 +41,4 @@ public abstract class Event implements Serializable{
         // HashSet of all active members of the event
     public HashMap<String, String> emailToDisplay = new HashMap<>();
         // HashMap relating a given email to a display name
-    public String eventDesc;
-        // Description of the event
-    //public MapView curEventMapView
-        // Reference to the map for this event
-    public String eventName;
-        // String storing the display name of this event
-    public int eventID;
-        // int storing the ID of the event.
-    public float radius;
-        // Radius of the event as set by a host
-    public String host;
 }
