+++ app/src/main/java/com/example/brianduffy/masevo/PublicEvent.java
@@ -1,13 +1,24 @@
 package com.example.brianduffy.masevo;
 
-import android.icu.text.SimpleDateFormat;
+import android.content.ContentValues;
 import android.os.Build;
 import android.support.annotation.RequiresApi;
 
+import java.io.BufferedReader;
+import java.io.DataOutputStream;
+import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.Serializable;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.sql.Date;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
+import java.util.Objects;
 
 /**
  * Created by Gabriel on 9/24/2017.
@@ -15,44 +26,222 @@ import java.util.HashSet;
 
 public class PublicEvent extends Event implements Serializable{
     // Use this to create a PublicEvent
+    private final String properties = "ID,Name,Description,StartTime,EndTime," +
+            "Latitude,Longitude,Radius,Host,List";
 
-    @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-    public PublicEvent(String eventName, String eventDesc, Date startTime, Date endTime,
+    private final String server_url = "http://webapp-171031005244.azurewebsites.net";
+
+    /**
+     * Used for call to server which do not require an event to be present
+     */
+    public PublicEvent() {};
+
+    public PublicEvent(String eventName, String eventDesc, Date startDate, Date endDate,
                        float latitude, float longitude, float radius, String creatorEmail)
     {
-        this.startTime = startTime;
-        this.endTime = endTime;
+        this.startDate = startDate;
+        this.endDate = endDate;
         this.location = new Location(latitude, longitude);
         this.eventName = eventName;
         this.eventDesc = eventDesc;
         this.radius = radius;
-        this.host = creatorEmail;
+        this.hostEmail = creatorEmail;
 
         this.hostList = new HashSet<>();
         this.attendeeList = new HashSet<>();
         this.activeList = new HashSet<>();
         this.emailToDisplay = new HashMap<>();
 
-        this.eventID = (eventName + eventDesc + startTime + endTime + latitude + longitude +
+        this.eventID = (eventName + eventDesc + startDate + endDate + latitude + longitude +
                 radius + Double.toString(Math.random())).hashCode();
-
-        // We now want to add this PublicEvent to the database using the proper call
-        ///////////////////////
-        //***DATABASE CALL***//
-        //***DATABASE CALL***//
-        //***DATABASE CALL***//
-        ///////////////////////
     }
 
     @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-    public void createPublicEvent () {
-        Server s = new Server();
-        s.createPublicEvent(eventID, eventName, eventDesc, startTime, endTime,
-                location.latitude, location.longitude, radius, host);
+    public void createPublicEvent() {
+        String methodName = "createPublicEvent";
+        String emailTrim = hostEmail.substring(0,hostEmail.indexOf("@"));
+
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("method",methodName);
+        contentValues.put("ID",Integer.toString(eventID));
+        contentValues.put("Name",eventName);
+        contentValues.put("Description",eventDesc);
+        contentValues.put("StartTime",Long.toString(startDate.getTime()));
+        contentValues.put("EndTime",Long.toString(endDate.getTime()));
+        contentValues.put("Latitude",Float.toString(location.latitude));
+        contentValues.put("Longitude",Float.toString(location.longitude));
+        contentValues.put("Radius",Float.toString(radius));
+        contentValues.put("Host",emailTrim);
+        String query = "";
+        for (Map.Entry e: contentValues.valueSet()) {
+            query += (e.getKey() + "=" + e.getValue() + "&");
+        }
+        query = query.substring(0, query.length() - 1);
+        final String fQuery = query;
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    byte[] postData = fQuery.getBytes(StandardCharsets.UTF_8);
+                    int postDataLength = postData.length;
+                    URL url = new URL(server_url);
+                    HttpURLConnection httpURLConnection = (HttpURLConnection)url.openConnection();
+                    httpURLConnection.setDoOutput(true);
+                    httpURLConnection.setInstanceFollowRedirects(false);
+                    httpURLConnection.setRequestMethod("POST");
+                    httpURLConnection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
+                    httpURLConnection.setRequestProperty("charset", "utf-8");
+                    httpURLConnection.setRequestProperty("Content-Length", Integer.toString(postDataLength));
+                    httpURLConnection.setUseCaches(false);
+                    try (DataOutputStream dataOutputStream = new DataOutputStream(httpURLConnection.getOutputStream())) {
+                        dataOutputStream.write(postData);
+                        dataOutputStream.flush();
+                    }
+                    int responseCode = httpURLConnection.getResponseCode();
+                    if (responseCode == 200) {
+                        String result = "";
+                        BufferedReader br = new BufferedReader(
+                                new InputStreamReader(httpURLConnection.getInputStream()));
+                        String output;
+                        while((output = br.readLine()) != null)
+                        {
+                            result += output;
+                        }
+                        System.out.println("Response message: " + result);
+                    }
+                } catch (MalformedURLException murle) {
+                    murle.printStackTrace();
+                    return;
+                } catch (IOException ioe) {
+                    ioe.printStackTrace();
+                    return;
+                }
+            }
+        }).start();
     }
 
-    public void deleteEvent(int ID) {
-        Server s = new Server();
-        s.deleteEvent(ID);
+    public ArrayList<PublicEvent> getPublicEvents () {
+        ThreadGetEvents threadGetEvents = new ThreadGetEvents();
+        new Thread(threadGetEvents).start();
+        return threadGetEvents.getReturnResult();
+    }
+
+    public void deleteEvent(int eventID) {
+        String methodName = "deleteEvent";
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("method",methodName);
+        contentValues.put("ID",Integer.toString(eventID));
+        String query = "";
+        for (Map.Entry e: contentValues.valueSet()) {
+            query += (e.getKey() + "=" + e.getValue() + "&");
+        }
+        query = query.substring(0, query.length() - 1);
+        final String fQuery = query;
+        new Thread(new Runnable() {
+            @RequiresApi(api = Build.VERSION_CODES.KITKAT)
+            @Override
+            public void run() {
+                try {
+                    byte[] postData = fQuery.getBytes(StandardCharsets.UTF_8);
+                    int postDataLength = postData.length;
+                    URL url = new URL(server_url);
+                    HttpURLConnection httpURLConnection = (HttpURLConnection)url.openConnection();
+                    httpURLConnection.setDoOutput(true);
+                    httpURLConnection.setInstanceFollowRedirects(false);
+                    httpURLConnection.setRequestMethod("POST");
+                    httpURLConnection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
+                    httpURLConnection.setRequestProperty("charset", "utf-8");
+                    httpURLConnection.setRequestProperty("Content-Length", Integer.toString(postDataLength));
+                    httpURLConnection.setUseCaches(false);
+                    try (DataOutputStream dataOutputStream = new DataOutputStream(httpURLConnection.getOutputStream())) {
+                        dataOutputStream.write(postData);
+                        dataOutputStream.flush();
+                    }
+                    int responseCode = httpURLConnection.getResponseCode();
+                    if (responseCode == 200) {
+                        String result = "";
+                        BufferedReader br = new BufferedReader(
+                                new InputStreamReader(httpURLConnection.getInputStream()));
+                        String output;
+                        while((output = br.readLine()) != null)
+                        {
+                            result += output;
+                        }
+                        System.out.println("Response message: " + result);
+                    }
+                } catch (MalformedURLException murle) {
+                    murle.printStackTrace();
+                    return;
+                } catch (IOException ioe) {
+                    ioe.printStackTrace();
+                    return;
+                }
+            }
+        }).start();
+    }
+
+    public void modifyEvent(int eventID, String eventName, String eventDesc, Date startDate, Date endDate,
+                       float latitude, float longitude, float radius, String hostEmail) {
+        String methodName = "modifyEvent";
+        String emailTrim = hostEmail.substring(0,hostEmail.indexOf("@"));
+
+        ContentValues contentValues = new ContentValues();
+        contentValues.put("method",methodName);
+        contentValues.put("ID",Integer.toString(eventID));
+        contentValues.put("Name",eventName);
+        contentValues.put("Description",eventDesc);
+        contentValues.put("StartTime",Long.toString(startDate.getTime()));
+        contentValues.put("EndTime",Long.toString(endDate.getTime()));
+        contentValues.put("Latitude",Float.toString(latitude));
+        contentValues.put("Longitude",Float.toString(longitude));
+        contentValues.put("Radius",Float.toString(radius));
+        contentValues.put("Host",emailTrim);
+        String query = "";
+        for (Map.Entry e: contentValues.valueSet()) {
+            query += (e.getKey() + "=" + e.getValue() + "&");
+        }
+        query = query.substring(0, query.length() - 1);
+        final String fQuery = query;
+        new Thread(new Runnable() {
+            @RequiresApi(api = Build.VERSION_CODES.KITKAT)
+            @Override
+            public void run() {
+                try {
+                    byte[] postData = fQuery.getBytes(StandardCharsets.UTF_8);
+                    int postDataLength = postData.length;
+                    URL url = new URL(server_url);
+                    HttpURLConnection httpURLConnection = (HttpURLConnection)url.openConnection();
+                    httpURLConnection.setDoOutput(true);
+                    httpURLConnection.setInstanceFollowRedirects(false);
+                    httpURLConnection.setRequestMethod("POST");
+                    httpURLConnection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
+                    httpURLConnection.setRequestProperty("charset", "utf-8");
+                    httpURLConnection.setRequestProperty("Content-Length", Integer.toString(postDataLength));
+                    httpURLConnection.setUseCaches(false);
+                    try (DataOutputStream dataOutputStream = new DataOutputStream(httpURLConnection.getOutputStream())) {
+                        dataOutputStream.write(postData);
+                        dataOutputStream.flush();
+                    }
+                    int responseCode = httpURLConnection.getResponseCode();
+                    if (responseCode == 200) {
+                        String result = "";
+                        BufferedReader br = new BufferedReader(
+                                new InputStreamReader(httpURLConnection.getInputStream()));
+                        String output;
+                        while((output = br.readLine()) != null)
+                        {
+                            result += output;
+                        }
+                        System.out.println("Response message: " + result);
+                    }
+                } catch (MalformedURLException murle) {
+                    murle.printStackTrace();
+                    return;
+                } catch (IOException ioe) {
+                    ioe.printStackTrace();
+                    return;
+                }
+            }
+        }).start();
     }
 }
