--- app/src/main/java/com/example/brianduffy/masevo/ResultSetter.java
@@ -1,11 +1,0 @@
-package com.example.brianduffy.masevo;
-
-import java.util.ArrayList;
-
-/**
- * Created by Josh on 11/3/2017.
- */
-
-public interface ResultSetter {
-    public void setResult(ArrayList<PublicEvent> result);
-}
